+++ CHANGELOG.md
@@ -22,9 +22,9 @@ Changes are grouped as follows
 - The method `client.functions.schedules.retrieve` now accepts the missing parameter `ignore_unknown_ids` as well
   as retrieving multiple schedules at once.
 - The method `client.functions.schedules.create` now supports creating using a `FunctionScheduleWrite` object.
-- 
+
 ### Changed
-- When creating a new function schedules without specifying `description`, the default values are now
+- When creating a new function schedule without specifying `description`, the default value is now
   correctly set to `None` instead of `""`.
 
 ## [7.54.7] - 2024-07-22